MOD: refactor bid section
