+++ src/pages/BidHNFT/components/BidSection/BidSection.less
@@ -1,0 +1,46 @@
+.bidSectionContainer {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: left;
+
+  .currentPrice {
+    width: 100%;
+  }
+
+  .currentPriceTitle {
+    color: rgb(112, 122, 131);
+    font-size: 16px;
+  }
+
+  .currentPriceValue {
+    font-size: 20px;
+    font-weight: 700;
+  }
+
+  .priceField {
+    width: 100%;
+    margin-top: 10px;
+    text-align: left;
+  }
+
+  .priceFieldTitle {
+    color: rgb(112, 122, 131);
+    font-size: 16px;
+  }
+
+  .tokenBalance {
+    text-align: right;
+  }
+
+  .bidBtnContainer {
+    width: 100%;
+    margin-top: 10px;
+    display: flex;
+    justify-content: center;
+
+    button {
+      max-width: 200px;
+    }
+  }
+}
